% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_probs.R
\name{convert_probs}
\alias{convert_probs}
\title{Convert probs object to new format.}
\usage{
convert_probs(object)
}
\arguments{
\item{object}{object of class \code{\link[qtl2geno]{calc_genoprob}} computed with old format}
}
\description{
Convert probs object to new format.
}
