% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pheno.R
\name{get_pheno}
\alias{get_pheno}
\title{Get phenotypes}
\usage{
get_pheno(phe, analyses_tbl, transform = TRUE)
}
\arguments{
\item{phe}{phenotypes in data frame}

\item{analyses_tbl}{table of analyses setups}

\item{transform}{transform phenotypes according to \code{analyses_tbl} if \code{TRUE}}
}
\value{
data frame of phenotypes
}
\description{
Get phenotypes using data frame of phenopypes filtered by \code{analyses_tbl}
}
\examples{
\dontrun{get_pheno(phe, analyses_tbl)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
