% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_probs.R
\name{modify_probs}
\alias{modify_probs}
\title{Modify probs object in new format.}
\usage{
modify_probs(object, newprob)
}
\arguments{
\item{object}{object of class \code{\link[qtl2geno]{calc_genoprob}}}

\item{newprob}{matrix of new LOD values}
}
\description{
Modify probs object in new format.
}
