% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nqrank.R
\name{nqrank}
\alias{nqrank}
\title{Quantile rank}
\usage{
nqrank(x, jitter = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{jitter}{jitter if \code{TRUE} (default \code{FALSE})}
}
\value{
transformed values
}
\description{
Quantile rank taken from R/qtl.
}
\examples{
dontrun(nqrank(x))

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
