% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno_tbl.R
\name{read_pheno_tbl}
\alias{read_pheno_tbl}
\title{Read phenotypes from RDS files}
\usage{
read_pheno_tbl(analyses_tbl, datapath)
}
\arguments{
\item{analyses_tbl}{table of analyses setups}

\item{datapath}{path to Derived Data}
}
\value{
data frame of phenotypes
}
\description{
Read phenotypes previously stored as RDS filtered by \code{analyses_tbl}
}
\examples{
\dontrun{read_pheno_tbl(analyses_tbl, datapath)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
