% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_probs.R
\name{read_probs}
\alias{read_probs}
\alias{read_probs36}
\title{Read genotype probability object from file}
\usage{
read_probs(chr = NULL, datapath)

read_probs36(chr_id, start_val, end_val, datapath)
}
\arguments{
\item{chr}{vector of chromosome identifiers}

\item{datapath}{name of folder with Derived Data}

\item{start_val, }{end_val start and end values in Mbp}

\item{chr}{vector of chromosome identifiers}

\item{datapath}{name of folder with Derived Data}
}
\value{
large object of class \code{\link[qtl2geno]{calc_genoprob}}

large object of class \code{\link[qtl2geno]{calc_genoprob}}
}
\description{
Uses readRDS to read object.

Uses readRDS to read object.
}
\examples{
\dontrun{read_probs(chr, datapath)}

\dontrun{read_probs36(chr, start_val, end_val, dirpath)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
